(#237)refactor: 회원 service interface 구현했습니다.
